requirements.txt
certifi==2025.8.3
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
fastparquet==2024.11.0
fonttools==4.60.0
fsspec==2025.9.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyarrow==21.0.0
pylatexenc==2.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==78.1.1
six==1.17.0
tzdata==2025.2
wincertstore==0.2